+++ RELEASE_NOTES.md
@@ -24,6 +24,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - Framing mosaics can now also be done with overlapping pixel count instead of an overlap percentage
 - Guider settings could only be set when being connected. Now these settings are also available in the setup screen of each guider without requiring a connection
 - When cloning a profile, the dock config for that profile will be cloned too
+- Simple Sequencer - "Cool Camera" toggled to on will also turn on dew heater when available as well as "Warm Camera" toggled to on will turn off the dew heater when enabled
 
 ## Features
 - Added interfaces and capabilities so that plugins can inject custom image meta data and image file patterns